+++ manager.js
@@ -32,7 +32,7 @@ manager.init({
 	tasks: {
 		prepare: {
 			postRun: ({ terminal }) => {
-				terminal.print(`Update Node version in package.json / .travis.yml / bitbucket-pipelines.yml`).spacer();
+				terminal.print(`Update Node version in package.json / bitbucket-pipelines.yml / .github/workflow/tests.yaml`).spacer();
 
 				const paths = require('./dist/node/helpers/paths');  // eslint-disable-line node/global-require
 				const today = Date.now();
@@ -55,11 +55,11 @@ manager.init({
 				fss.writeJson(packageFile, packageData, { space: 2 });
 
 
-				//-- .travis.yml
-				const travisFile = `${paths.root}/.travis.yml`;
-				const travisData = fss.readYaml(travisFile);
-				travisData.node_js = ['node', ...lts];  // eslint-disable-line camelcase
-				fss.writeYaml(travisFile, travisData);
+				//-- .github/workflow/tests.yaml
+				const githubActionsFile = `${paths.root}/.github/workflow/tests.yaml`;
+				const githubActionsData = fss.readYaml(githubActionsFile);
+				githubActionsData.jobs.build.strategy.matrix.node_version = [...lts];  // eslint-disable-line camelcase
+				fss.writeYaml(githubActionsFile, githubActionsData);
 
 
 				//-- bitbucket-pipelines.yml