+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- Add cache flag for Bitbucket Pipelines cache (#15)
+
+### Changed
+- Migrate from TravisCI to GitHub Actions (#19)